Chagne Limit

+++ mm/internal_mem.h
@@ -1,4 +1,4 @@
 #ifndef __MM_INTERNAL_MEM_H
 #define __MM_INTERNAL_MEM_H
-extern unsigned int __limit__memory__;
+extern unsigned long __limit__memory__;
 #endif	/* __MM_INTERNAL_H */
